tests: add record fixtures
* Fixes alembic scripts.
* Creates records fixtures for agents and concepts.
* Adds a new create record test utils function.
