+++ rero_mef/agents/__init__.py
@@ -23,8 +23,8 @@ from .mef.api import AgentMefIndexer, AgentMefRecord, AgentMefSearch
 from .rero.api import AgentReroIndexer, AgentReroRecord, AgentReroSearch
 from .viaf.api import AgentViafIndexer, AgentViafRecord, AgentViafSearch
 
-__all__ = (AgentGndIndexer, AgentGndRecord, AgentGndSearch,
-           AgentIdrefIndexer, AgentIdrefRecord, AgentIdrefSearch,
-           AgentMefIndexer, AgentMefRecord, AgentMefSearch,
-           AgentReroIndexer, AgentReroRecord, AgentReroSearch,
-           AgentViafIndexer, AgentViafRecord, AgentViafSearch)
+__all__ = ('AgentGndIndexer', 'AgentGndRecord', 'AgentGndSearch',
+           'AgentIdrefIndexer', 'AgentIdrefRecord', 'AgentIdrefSearch',
+           'AgentMefIndexer', 'AgentMefRecord', 'AgentMefSearch',
+           'AgentReroIndexer', 'AgentReroRecord', 'AgentReroSearch',
+           'AgentViafIndexer', 'AgentViafRecord', 'AgentViafSearch')
