+++ rero_mef/concepts/__init__.py
@@ -20,5 +20,5 @@
 from .mef.api import ConceptMefIndexer, ConceptMefRecord, ConceptMefSearch
 from .rero.api import ConceptReroIndexer, ConceptReroRecord, ConceptReroSearch
 
-__all__ = (ConceptMefIndexer, ConceptMefRecord, ConceptMefSearch,
-           ConceptReroIndexer, ConceptReroRecord, ConceptReroSearch)
+__all__ = ('ConceptMefIndexer', 'ConceptMefRecord', 'ConceptMefSearch',
+           'ConceptReroIndexer', 'ConceptReroRecord', 'ConceptReroSearch')
