+++ tests/api/test_agents_idref_rest.py
@@ -15,27 +15,17 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-"""Test REST API GND."""
+"""Test REST API IDREF."""
 
 import json
 
 from flask import url_for
 
-from rero_mef.agents.gnd.api import AgentGndRecord
 
-
-def test_view_agents_gnd(client, agent_gnd_record):
-    """Test redirect GND."""
-
-    AgentGndRecord.create(
-        data=agent_gnd_record,
-        delete_pid=False,
-        dbcommit=True,
-        reindex=True,
-    )
-    AgentGndRecord.flush_indexes()
-    pid = agent_gnd_record.get('pid')
-    url = url_for('api_agents_gnd.redirect_list')
+def test_view_agents_idref(client, agent_idref_record):
+    """Test redirect IDREF."""
+    pid = agent_idref_record.get('pid')
+    url = url_for('api_agents_idref.redirect_list')
     res = client.get(url)
     assert res.status_code == 308
     res = client.get(url, follow_redirects=True)
@@ -49,7 +39,7 @@ def test_view_agents_gnd(client, agent_gnd_record):
         'deleted': {'doc_count': 0}
     }
 
-    url = url_for('api_agents_gnd.redirect_item', pid=pid)
+    url = url_for('api_agents_idref.redirect_item', pid=pid)
     res = client.get(url)
     assert res.status_code == 308
     res = client.get(url, follow_redirects=True)
@@ -58,7 +48,7 @@ def test_view_agents_gnd(client, agent_gnd_record):
     assert data.get('id') == pid
     assert data.get('metadata', {}).get('pid') == pid
 
-    url = url_for('api_agents_gnd.redirect_item', pid='WRONG')
+    url = url_for('api_agents_idref.redirect_item', pid='WRONG')
     res = client.get(url, follow_redirects=True)
     assert res.status_code == 404
     assert json.loads(res.get_data(as_text=True)) == {
