+++ tests/ui/agents/test_ref_resolver.py
