+++ tests/ui/test_cli.py
